+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Create release command line utility [#6](https://github.com/stevewest/changelog/issues/6).
+- Create convert command line utility [#6](https://github.com/stevewest/changelog/issues/6).
 
 ### Fixed
 - Links are no longer parsed as a description.